# services/rider/swagger.yaml
openapi: '3.0.3'
info:
  title: MealStack Rider Service API
  description: Rider microservice for MealStack food delivery platform
  version: 1.0.0
  contact:
    name: MealStack Team
servers:
  - url: http://localhost:3004
    description: Local development
...
paths:
...
  /api/rider/location:
    patch:
      tags: [Location]
      summary: Update rider GPS location
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat:
                  type: number
                  example: 6.9271
                lng:
                  type: number
                  example: 79.8612
      responses:
        '200':
          description: Location updated
    get:
      tags: [Location]
      summary: Get rider current location
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current location
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...